api/requirements.txt
SQLAlchemy==1.4.46
pymysql==1.0.2
Flask_RESTful==0.3.9
Flask==2.2.2
Flask_SQLAlchemy==3.0.2
Flask_Marshmallow==0.14.0
flask_cors==4.0.0
regex==2023.3.23
Werkzeug==2.2.2
Gunicorn
pyJWT==2.8.0